update child classes about abstract list class
